+++ commands/__init__.py
@@ -1,0 +1,1 @@
+__all__ = ["hi_contrib", "api_commands", "wiki_search"]
