+++ commands/scripts/moderate_words.txt
