+++ commands/scripts/moderator.py
